Move TaskyTableModel to taskytable package
