+++ src/ui/taskytable/TaskyTableModel.java
@@ -1,4 +1,4 @@
-package ui;
+package ui.taskytable;
 
 import javax.swing.table.AbstractTableModel;
